control_turtlebot/requirements.txt
# TurtleBot Controller - Python Dependencies

# Core dependencies
pygame>=2.0.0
pyyaml>=5.4.0

# Networking
websockets>=10.0

# Scientific
numpy>=1.20.0
